+++ demisto_sdk/commands/upload/tests/uploader_test.py
@@ -954,7 +954,7 @@ class TestZippedPackUpload:
             - call to upload command
         Then:
             - validate the upload_content_packs in the api client was called correct
-              and the pack verification ws turned on and off
+              and the skip_verify arg is "true"
         """
         # prepare
         mock_api_client(mocker)
@@ -970,6 +970,36 @@ class TestZippedPackUpload:
 
         assert str(uploaded_file_path) == input
         assert skip_value == 'true'
+
+    @pytest.mark.parametrize(argnames='input', argvalues=[TEST_PACK_ZIP, CONTENT_PACKS_ZIP])
+    def test_upload_without_skip_verify(self, mocker, input):
+        """
+        Given:
+            - zipped pack or zip of pack zips to upload
+        When:
+            - call to upload command
+        Then:
+            - validate the upload_content_packs in the api client was called correct
+              and the skip_verify arg is None
+        """
+        # prepare
+        mock_api_client(mocker)
+        mocker.patch.object(API_CLIENT, 'upload_content_packs')
+        mocker.patch.object(tools, 'update_server_configuration', return_value=(None, None, {}))
+        mocker.patch.object(Pack, 'is_server_version_ge', return_value=False)
+        mocker.patch.object(Uploader, 'notify_user_should_override_packs', return_value=True)
+
+        # run
+        click.Context(command=upload).invoke(upload, input=input)
+        uploaded_file_path = API_CLIENT.upload_content_packs.call_args[1]['file']
+
+        assert str(uploaded_file_path) == input
+
+        try:
+            skip_value = API_CLIENT.upload_content_packs.call_args[1]['skip_verify']
+        except:
+            skip_value = None
+        assert not skip_value
 
 
 def exception_raiser(**kwargs):